requirements.txt
asgiref==3.5.2
backports.zoneinfo ==0.2.1
Django==4.0.5
django-model-utils==4.2.0
djangorestframework==3.13.1
pip==19.2.3
pytz==2022.1
setuptools==41.2.0
sqlparse==0.4.2
tzdata==2022.1
